Initial commit. Some basic functions and tests.
Blowfish Encrypt and Decrypt methods are fully fleshed out. So is the
pandoraCall method, which sends an HTTP request and decodes the JSON
response. Two wrappers around pandoraCall as well that provide either
TLS or Blowfish encryption.

Rudimentary tests for the encrypt and decrypt methods. Also for the
PartnerLogin method though that method is not finished nor is the test.